+++ app/checkout/page.tsx
@@ -1,9 +1,5 @@
-'use client'
 import { Metadata } from 'next';
-import {Elements} from '@stripe/react-stripe-js';
-import {loadStripe} from '@stripe/stripe-js';
-import CheckoutForm from '@/components/Checkout/CheckoutForm';
-
+import StripeWrapper from '@/components/Checkout/StripeWrapper';
 
 export const metadata: Metadata = {
   title: 'DNTrademark Admin - Global Trademark Notification Platform',
@@ -11,21 +7,8 @@ export const metadata: Metadata = {
   // other metadata
 };
 
-// Docs:: https://stripe.com/docs/payments/accept-a-payment-charges?client=react
-
-// Make sure to call `loadStripe` outside of a component's render to avoid
-// recreating the `Stripe` object on every render.
-const stripePromise = loadStripe('pk_test_TYooMQauvdEDq54NiTphI7jx');
-
 export default function App() {
-  const options = {
-    // passing the client secret obtained from the server
-    clientSecret: '{{CLIENT_SECRET}}',
-  };
-
   return (
-    <Elements stripe={stripePromise}>
-      <CheckoutForm />
-    </Elements>
+      <StripeWrapper />
   );
 };
